Update.
1999-10-12  Andreas Jaeger  <aj@suse.de>

	* stdlib/tst-environ.c: Include <string.h> for strcpy declaration.

	* math/basic-test.c: Fix typo.

	* locale/programs/xstrdup.c: Include string.h for glibc
	compilation to get string prototypes.

	* sysdeps/unix/sysv/linux/alpha/sys/ucontext.h: Add definition of
	gregset_t and fpregset_t.

	* sysdeps/generic/s_nexttowardf.c: Add parenthesis to shut up gcc
	warnings.  Correct value for x == 0.

	* argp/argp-help.c: Remove broken definition of flockfile and
	funlockfile.

1999-10-12  Andreas Schwab  <schwab@suse.de>

	* sysdeps/unix/sysv/linux/alpha/bits/sigaction.h: Update to match
	generic Linux version.

1999-10-12  Ulrich Drepper  <drepper@cygnus.com>

	* sysdeps/unix/sysv/linux/bits/sigaction.h: Remove K&R compatibility.

+++ linuxthreads/internals.h
@@ -358,9 +358,9 @@ extern void __pthread_kill_other_threads_np __P ((void));
 extern int __libc_close (int fd);
 extern int __libc_nanosleep (const struct timespec *requested_time,
 			     struct timespec *remaining);
-extern int __libc_read (int fd, void *buf, size_t count);
+extern ssize_t __libc_read (int fd, void *buf, size_t count);
 extern pid_t __libc_waitpid (pid_t pid, int *stat_loc, int options);
-extern int __libc_write (int fd, const void *buf, size_t count);
+extern ssize_t __libc_write (int fd, const void *buf, size_t count);
 
 /* Prototypes for some of the new semaphore functions.  */
 extern int __new_sem_post (sem_t * sem);
